Missing includes, boyscout and clang compilation issues fixed

+++ source/runtime/ppp_engine/private/util/hash.h
@@ -1,11 +1,13 @@
 #pragma once
 
+#include "util/types.h"
+
 namespace ppp
 {
     namespace utils
     {
         template <typename T>
-        inline size_t hash_combine(size_t seed, const T& v)
+        inline u64 hash_combine(u64 seed, const T& v)
         {
             std::hash<T> hasher;
             seed ^= hasher(v) + 0x9e3779b9 + (seed<<6) + (seed>>2);
